refactor(pallet-communities): move BenchmarkHelper to types module
